Rename person package to note
