+++ src/main/java/seedu/address/logic/commands/DeleteNoteCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 
 /**
  * Deletes a lecture note by index number.
