+++ src/main/java/seedu/address/model/Model.java
@@ -5,7 +5,7 @@ import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 
 /**
  * The API of the Model component.
