+++ src/main/java/seedu/address/model/note/TitleContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.note;
 
 import java.util.List;
 import java.util.function.Predicate;
