+++ src/main/java/seedu/address/model/note/exceptions/DuplicateTitleException.java
@@ -1,0 +1,10 @@
+package seedu.address.model.note.exceptions;
+
+/**
+ * Signals that the operation will result in lecture notes with duplicate titles.
+ */
+public class DuplicateTitleException extends RuntimeException {
+    public DuplicateTitleException() {
+        super("Operation would result in lecture notes with duplicate titles");
+    }
+}
