+++ src/main/java/seedu/address/model/note/exceptions/NoteNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.note.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified lecture note.
