+++ src/main/java/seedu/address/ui/NoteCard.java
@@ -4,7 +4,7 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Label;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 
 /**
  * An UI component that displays information of a {@code Note}.
