+++ src/main/java/seedu/address/ui/NoteListPanel.java
@@ -8,7 +8,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 
 /**
  * Panel containing the list of lecture notes.
