+++ src/test/java/seedu/address/logic/commands/AddNoteCommandIntegrationTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 import seedu.address.testutil.PersonBuilder;
 
 /**
