+++ src/test/java/seedu/address/logic/commands/AddNoteCommandTest.java
@@ -20,7 +20,7 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.person.Note;
+import seedu.address.model.note.Note;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddNoteCommandTest {
