+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -22,8 +22,8 @@ import seedu.address.logic.commands.FindNoteCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListNoteCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Note;
-import seedu.address.model.person.TitleContainsKeywordsPredicate;
+import seedu.address.model.note.Note;
+import seedu.address.model.note.TitleContainsKeywordsPredicate;
 import seedu.address.testutil.EditNoteDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
 import seedu.address.testutil.PersonUtil;
