+++ src/test/java/seedu/address/logic/parser/ParserUtilTest.java
@@ -8,8 +8,8 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Content;
-import seedu.address.model.person.Title;
+import seedu.address.model.note.Content;
+import seedu.address.model.note.Title;
 
 public class ParserUtilTest {
     private static final String INVALID_TITLE = "a\nb";
