+++ src/test/java/seedu/address/testutil/EditNoteDescriptorBuilder.java
@@ -2,9 +2,9 @@ package seedu.address.testutil;
 
 import seedu.address.logic.commands.EditNoteCommand;
 import seedu.address.logic.commands.EditNoteCommand.EditNoteDescriptor;
-import seedu.address.model.person.Content;
-import seedu.address.model.person.Note;
-import seedu.address.model.person.Title;
+import seedu.address.model.note.Content;
+import seedu.address.model.note.Note;
+import seedu.address.model.note.Title;
 
 /**
  * A utility class to help with building EditNoteDescriptor objects.
